+++ Sources/CorePayments/Networking/GraphQL/GraphQLQuery.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+//// :nodoc: Protocol for requests made to GraphQL endpoints
 public protocol GraphQLQuery {
     associatedtype VariablesType: Codable
     var query: String { get }