+++ Sources/CorePayments/Networking/GraphQL/GraphQLQueryResponse.swift
@@ -1,3 +1,4 @@
+//// :nodoc: Struct to handle nested response from GraphQL response
 public struct GraphQLQueryResponse<T: Codable>: Codable {
 
     public let data: T?