+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## unreleased
 * CardPayments
-    * Add `vault`property in `CardRequest`, this type contains a `customerID`, if a customer ID is not provided one will be generated  
+  * Add `vault`property in `CardRequest`, this type contains a `customerID`, if a customer ID is not provided one will be generated  
 * PayPalNativePayments
   * Bump `PayPalCheckout` to `1.0.0`