+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 ## unreleased
 * PayPalNativePayments
   * Bump `PayPalCheckout` to `1.0.0`
+* CardPayments
+  * Add `vault` method 
+  * Add `CardVaultRequest` and `CardVaultResult` types for interacting with `vault` method
+  * Add `CardVaultDelegate` protocol to receive success and failure results
+  * Add `CardVaultDelegate` property to `CardClient`
   
 ## 0.0.9 (2023-06-23)
 * Breaking Changes